dcn-worker/requirements.txt
requests
psutil
Pillow
beautifulsoup4
openai-whisper
